backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
django-nested-admin==4.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
gunicorn==21.2.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psycopg==3.1.14
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
